requirements.txt
streamlit>=1.32.0
langchain>=0.1.20
langchain-community>=0.0.38
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.3
langgraph>=0.2.0
chromadb>=0.4.24
sentence-transformers>=2.7.0
groq>=0.5.0
pypdf>=4.2.0
reportlab>=4.1.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
schedule>=1.2.1
pydantic>=2.7.0
